requirements.txt
langchain==0.3.25
openai==1.84.0
langchain-openai==0.3.21
playwright==1.52.0
python-dotenv==1.1.0
feedparser==6.0.11
Pillow==10.2.0
requests==2.32.3
pymysql==1.1.1
sshtunnel==0.4.0